added response-helper file to use EventEmitter for respondingto clients
