+++ server.js
@@ -8,7 +8,7 @@ app.use(express.json());
 import apiRouter from "./server/routes/api.js";
 
 //APP TURNED OFF TO SAVE MONEY DURING TESTING
-// app.use("/api", apiRouter);
+app.use("/api", apiRouter);
 
 app.use((req, res) => res.sendStatus(404));
